Use actor-protected `uint64` instead of atomics

+++ cmd/pineconesim/simulator/state.go
@@ -29,12 +29,12 @@ type RootAnnouncement struct {
 
 type PeerBandwidthUsage struct {
 	Protocol struct {
-		Rx uint32
-		Tx uint32
+		Rx uint64
+		Tx uint64
 	}
 	Overlay struct {
-		Rx uint32
-		Tx uint32
+		Rx uint64
+		Tx uint64
 	}
 }
